Ignore techs that don't exist
Sometimes techs are switched off, either at the 'techs' level or at all 'locations', leading to them not being in the model. The definitive list of techs in the model, and their corresponding carriers, are defined in the sets `loc_tech_carriers_[prod/con]`. This commit processes these sets directly to build the `techs` and `carriers` to be connected in the model graph.